Update emby.json
adding the https port per: https://support.emby.media/support/solutions/articles/44001159258-advanced-menu to the NAT_Forwards. Changing the name to reflect the difference between this and the other port.

I would like to hear comments about making this the beta channel plug-in to make it usefully different from the other emby plug-in?