restoring artefact
if the branch command works this should enable pushing this back to iX and have both old python 2 version work on artefact branch of master and when the python3 PR is accepted to this .json it should work with python3 branch of artefact. 

Once this manifest is live then can do PR to remove branch tag at same time as merging artefact branches.