Update jellyfin.json to 13.3-RELEASE
To match the contents of https://github.com/spz2k9/iocage-plugin-jellyfin/pull/12

For the same reason as stated in that PR: the port of jellyfin doesn't work except on 13.3 or 14+. Obviously you will need to have manually installed TrueNAS 13.3 for this to work, but this plugin doesn't work on TrueNAS 13.0 as-is anyway. Before TrueNAS CORE goes the way of the dodo, it'd be nice for this plugin to work.